+++ internal/sqlc/models.go
@@ -220,7 +220,7 @@ type DocumentClient struct {
 }
 
 type Format struct {
-	FormatID        int64            `json:"format_id"`
+	FormatID        int32            `json:"format_id"`
 	UpdatedFormatID pgtype.Int4      `json:"updated_format_id"`
 	AsesorID        int32            `json:"asesor_id"`
 	FormatName      string           `json:"format_name"`
